+++ hurdlethon/models/__init__.py
@@ -1,0 +1,7 @@
+# __init__.py of models
+from .question import Question
+from .answer import Answer
+from .lecture import Lecture
+from .user import User
+
+__all__ = ["Question", "Answer", "Lecture", "User"]
