+++ hurdlethon/serializers/lecture.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from ..models.lecture import Lecture
+
+class LectureSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Lecture
+        fields = ['LECTURE_ID', 'NAME', 'LECTURE_DAY', 'LECTURE_TIME']
