+++ hurdlethon/serializers/login.py
@@ -1,0 +1,6 @@
+from rest_framework import serializers
+
+
+class LoginSerializer(serializers.Serializer):
+    username = serializers.CharField()
+    password = serializers.CharField(write_only=True)
